pipeline/requirements.txt
appdirs==1.4.4
attrs==22.2.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///Users/cbousseau/work/recipes/ci_py311/certifi_1677903144932/work/certifi
cffi @ file:///Users/cbousseau/work/recipes/ci_py311/cffi_1677903595907/work
chardet==5.1.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.0.7
cryptography @ file:///Users/cbousseau/work/recipes/ci_py311_2/cryptography_1678994456782/work
cycler==0.11.0
fonttools==4.39.2
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
idna @ file:///Users/cbousseau/work/recipes/ci_py311/idna_1677906072337/work
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==3.3.5
matplotlib==3.7.1
numpy @ file:///Users/cbousseau/work/recipes/ci_py311/numpy_and_numpy_base_1677921358445/work
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e946luvhc3/croot/packaging_1678965323926/work
pandas==1.5.3
parse==1.19.0
Pillow==9.4.0
pluggy @ file:///Users/cbousseau/work/recipes/ci_py311/pluggy_1677906980825/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
protobuf==4.22.1
py @ file:///opt/conda/conda-bld/py_1644396412707/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ddecjfxr7g/croot/pyopenssl_1678965309681/work
pyparsing==3.0.9
PySocks @ file:///Users/cbousseau/work/recipes/ci_py311/pysocks_1677906386870/work
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
scikit-learn==1.1.3
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tomli @ file:///Users/cbousseau/work/recipes/ci_py311/tomli_1677905024737/work
tqdm==4.65.0
urllib3==1.26.15
xgboost==1.7.4
